bat file for toggling built-in keyboard
super easy to use. Run the file as administrator in order for the changes to take effect. Type 1 or 2 into the command line, depending on whether you would like to enable or disable the keyboard. Hit enter, and restart your laptop. It's as easy as that. In my case, I needed to disable my built-in keyboard since it's a bit busted, and when I set my bluetooth keyboard on my laptop it tends to randomly hit the keys it's resting on if we hit a bump (I write my book on the road while my coworkers drive). There you have it. Super simple tool really, it basically just runs the [sc config i8042prt] command for you so you don't have to go through the trouble of doing it yourself, but also with the enabler it just streamlines the whole process.